Refactor: split TFConfigurationMeta to single package
