+++ api/types/terraform.go
@@ -14,3 +14,10 @@ const (
 	// ConfigurationRemote means HCL stores in a remote git repository
 	ConfigurationRemote ConfigurationType = "Remote"
 )
+
+// GitRef specifies the git reference
+type GitRef struct {
+	Branch string `json:"branch,omitempty"`
+	Tag    string `json:"tag,omitempty"`
+	Commit string `json:"commit,omitempty"`
+}
