+++ controllers/configuration_controller.go
@@ -19,12 +19,11 @@ package controllers
 import (
 	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
+	"github.com/oam-dev/terraform-controller/controllers/process"
+	"github.com/oam-dev/terraform-controller/controllers/util"
 	"math"
 	"os"
-	"path/filepath"
-	"reflect"
 	"strconv"
 	"time"
 
@@ -43,11 +42,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
 	"github.com/oam-dev/terraform-controller/api/types"
-	crossplane "github.com/oam-dev/terraform-controller/api/types/crossplane-runtime"
-	"github.com/oam-dev/terraform-controller/api/v1beta1"
 	"github.com/oam-dev/terraform-controller/api/v1beta2"
 	tfcfg "github.com/oam-dev/terraform-controller/controllers/configuration"
-	"github.com/oam-dev/terraform-controller/controllers/configuration/backend"
 	"github.com/oam-dev/terraform-controller/controllers/features"
 	"github.com/oam-dev/terraform-controller/controllers/provider"
 	"github.com/oam-dev/terraform-controller/controllers/terraform"
@@ -55,16 +51,6 @@ import (
 
 const (
 	defaultNamespace = "default"
-	// WorkingVolumeMountPath is the mount path for working volume
-	WorkingVolumeMountPath = "/data"
-	// InputTFConfigurationVolumeName is the volume name for input Terraform Configuration
-	InputTFConfigurationVolumeName = "tf-input-configuration"
-	// BackendVolumeName is the volume name for Terraform backend
-	BackendVolumeName = "tf-backend"
-	// InputTFConfigurationVolumeMountPath is the volume mount path for input Terraform Configuration
-	InputTFConfigurationVolumeMountPath = "/opt/tf-configuration"
-	// BackendVolumeMountPath is the volume mount path for Terraform backend
-	BackendVolumeMountPath = "/opt/tf-backend"
 	// terraformContainerName is the name of the container that executes the terraform in the pod
 	terraformContainerName     = "terraform-executor"
 	terraformInitContainerName = "terraform-init"
@@ -140,7 +126,7 @@ func (r *ConfigurationReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 		return ctrl.Result{}, client.IgnoreNotFound(err)
 	}
 
-	meta := initTFConfigurationMeta(req, configuration, r.Client)
+	meta := process.New(req, configuration, r.Client)
 	if r.ControllerNamespace != "" {
 		uid := string(configuration.GetUID())
 		// @step: since we are using a single namespace to run these, we must ensure the names
@@ -178,13 +164,13 @@ func (r *ConfigurationReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 		_, err := terraform.GetTerraformStatus(ctx, meta.ControllerNamespace, meta.DestroyJobName, terraformContainerName, terraformInitContainerName)
 		if err != nil {
 			klog.ErrorS(err, "Terraform destroy failed")
-			if updateErr := meta.updateDestroyStatus(ctx, r.Client, types.ConfigurationDestroyFailed, err.Error()); updateErr != nil {
+			if updateErr := meta.UpdateDestroyStatus(ctx, r.Client, types.ConfigurationDestroyFailed, err.Error()); updateErr != nil {
 				return ctrl.Result{}, updateErr
 			}
 		}
 
 		// If no tfState has been generated, then perform a quick cleanup without dispatching destroying job.
-		if meta.isTFStateGenerated(ctx) {
+		if meta.IsTFStateGenerated(ctx) {
 			if err := r.terraformDestroy(ctx, configuration, meta); err != nil {
 				if err.Error() == types.MessageDestroyJobNotCompleted {
 					return ctrl.Result{RequeueAfter: 3 * time.Second}, nil
@@ -212,9 +198,9 @@ func (r *ConfigurationReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 	}
 
 	var tfExecutionJob = &batchv1.Job{}
-	if err := meta.getApplyJob(ctx, r.Client, tfExecutionJob); err == nil {
+	if err := meta.GetApplyJob(ctx, r.Client, tfExecutionJob); err == nil {
 		if !meta.EnvChanged && !meta.ConfigurationChanged && tfExecutionJob.Status.Succeeded == int32(1) {
-			err = meta.updateApplyStatus(ctx, r.Client, types.Available, types.MessageCloudResourceDeployed)
+			err = meta.UpdateApplyStatus(ctx, r.Client, types.Available, types.MessageCloudResourceDeployed)
 			return ctrl.Result{}, err
 		}
 	}
@@ -230,7 +216,7 @@ func (r *ConfigurationReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 	state, err := terraform.GetTerraformStatus(ctx, meta.ControllerNamespace, meta.ApplyJobName, terraformContainerName, terraformInitContainerName)
 	if err != nil {
 		klog.ErrorS(err, "Terraform apply failed")
-		if updateErr := meta.updateApplyStatus(ctx, r.Client, state, err.Error()); updateErr != nil {
+		if updateErr := meta.UpdateApplyStatus(ctx, r.Client, state, err.Error()); updateErr != nil {
 			return ctrl.Result{}, updateErr
 		}
 
@@ -240,158 +226,27 @@ func (r *ConfigurationReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 	return ctrl.Result{}, nil
 }
 
-// LegacySubResources if user specify ControllerNamespace when re-staring controller, there are some sub-resources like Secret
-// and ConfigMap that are in the namespace of the Configuration. We need to GC these sub-resources when Configuration is deleted.
-type LegacySubResources struct {
-	// Namespace is the namespace of the Configuration, also the namespace of the sub-resources.
-	Namespace           string
-	ApplyJobName        string
-	DestroyJobName      string
-	ConfigurationCMName string
-	VariableSecretName  string
-}
-
-type ResourceQuota struct {
-	ResourcesLimitsCPU              string
-	ResourcesLimitsCPUQuantity      resource.Quantity
-	ResourcesLimitsMemory           string
-	ResourcesLimitsMemoryQuantity   resource.Quantity
-	ResourcesRequestsCPU            string
-	ResourcesRequestsCPUQuantity    resource.Quantity
-	ResourcesRequestsMemory         string
-	ResourcesRequestsMemoryQuantity resource.Quantity
-}
-
-// TFConfigurationMeta is all the metadata of a Configuration
-type TFConfigurationMeta struct {
-	Name                                         string
-	Namespace                                    string
-	ControllerNamespace                          string
-	ConfigurationType                            types.ConfigurationType
-	CompleteConfiguration                        string
-	RemoteGit                                    string
-	RemoteGitPath                                string
-	ConfigurationChanged                         bool
-	EnvChanged                                   bool
-	ConfigurationCMName                          string
-	ApplyJobName                                 string
-	DestroyJobName                               string
-	Envs                                         []v1.EnvVar
-	ProviderReference                            *crossplane.Reference
-	VariableSecretName                           string
-	VariableSecretData                           map[string][]byte
-	DeleteResource                               bool
-	Region                                       string
-	Credentials                                  map[string]string
-	JobEnv                                       map[string]interface{}
-	GitCredentialsSecretReference                *v1.SecretReference
-	TerraformCredentialsSecretReference          *v1.SecretReference
-	TerraformRCConfigMapReference                *v1.SecretReference
-	TerraformCredentialsHelperConfigMapReference *v1.SecretReference
-
-	Backend backend.Backend
-	// JobNodeSelector Expose the node selector of job to the controller level
-	JobNodeSelector map[string]string
-
-	// TerraformImage is the Terraform image which can run `terraform init/plan/apply`
-	TerraformImage string
-	BusyboxImage   string
-	GitImage       string
-
-	// BackoffLimit specifies the number of retries to mark the Job as failed
-	BackoffLimit int32
-
-	// ResourceQuota series Variables are for Setting Compute Resources required by this container
-	ResourceQuota ResourceQuota
-
-	LegacySubResources    LegacySubResources
-	ControllerNSSpecified bool
-
-	K8sClient client.Client
-}
-
-func initTFConfigurationMeta(req ctrl.Request, configuration v1beta2.Configuration, k8sClient client.Client) *TFConfigurationMeta {
-	var meta = &TFConfigurationMeta{
-		ControllerNamespace: req.Namespace,
-		Namespace:           req.Namespace,
-		Name:                req.Name,
-		ConfigurationCMName: fmt.Sprintf(TFInputConfigMapName, req.Name),
-		VariableSecretName:  fmt.Sprintf(TFVariableSecret, req.Name),
-		ApplyJobName:        req.Name + "-" + string(TerraformApply),
-		DestroyJobName:      req.Name + "-" + string(TerraformDestroy),
-		K8sClient:           k8sClient,
-	}
-
-	jobNodeSelectorStr := os.Getenv("JOB_NODE_SELECTOR")
-	if jobNodeSelectorStr != "" {
-		err := json.Unmarshal([]byte(jobNodeSelectorStr), &meta.JobNodeSelector)
-		if err != nil {
-			klog.Warningf("the value of JobNodeSelector is not a json string ", err)
-		}
-	}
-
-	// githubBlocked mark whether GitHub is blocked in the cluster
-	githubBlockedStr := os.Getenv("GITHUB_BLOCKED")
-	if githubBlockedStr == "" {
-		githubBlockedStr = "false"
-	}
-
-	meta.RemoteGit = tfcfg.ReplaceTerraformSource(configuration.Spec.Remote, githubBlockedStr)
-	if configuration.Spec.DeleteResource != nil {
-		meta.DeleteResource = *configuration.Spec.DeleteResource
-	} else {
-		meta.DeleteResource = true
-	}
-	if configuration.Spec.Path == "" {
-		meta.RemoteGitPath = "."
-	} else {
-		meta.RemoteGitPath = configuration.Spec.Path
-	}
-
-	if !configuration.Spec.InlineCredentials {
-		meta.ProviderReference = tfcfg.GetProviderNamespacedName(configuration)
-	}
-
-	if configuration.Spec.GitCredentialsSecretReference != nil {
-		meta.GitCredentialsSecretReference = configuration.Spec.GitCredentialsSecretReference
-	}
-
-	if configuration.Spec.TerraformCredentialsSecretReference != nil {
-		meta.TerraformCredentialsSecretReference = configuration.Spec.TerraformCredentialsSecretReference
-	}
-
-	if configuration.Spec.TerraformRCConfigMapReference != nil {
-		meta.TerraformRCConfigMapReference = configuration.Spec.TerraformRCConfigMapReference
-	}
-
-	if configuration.Spec.TerraformCredentialsHelperConfigMapReference != nil {
-		meta.TerraformCredentialsHelperConfigMapReference = configuration.Spec.TerraformCredentialsHelperConfigMapReference
-	}
-
-	return meta
-}
-
-func (r *ConfigurationReconciler) terraformApply(ctx context.Context, configuration v1beta2.Configuration, meta *TFConfigurationMeta) error {
+func (r *ConfigurationReconciler) terraformApply(ctx context.Context, configuration v1beta2.Configuration, meta *process.TFConfigurationMeta) error {
 
 	var (
 		k8sClient      = r.Client
 		tfExecutionJob batchv1.Job
 	)
 
-	if err := meta.getApplyJob(ctx, k8sClient, &tfExecutionJob); err != nil {
+	if err := meta.GetApplyJob(ctx, k8sClient, &tfExecutionJob); err != nil {
 		if kerrors.IsNotFound(err) {
-			return meta.assembleAndTriggerJob(ctx, k8sClient, TerraformApply)
+			return meta.AssembleAndTriggerJob(ctx, k8sClient, process.TerraformApply)
 		}
 	}
 	klog.InfoS("terraform apply job", "Namespace", tfExecutionJob.Namespace, "Name", tfExecutionJob.Name)
 
-	if err := meta.updateTerraformJobIfNeeded(ctx, k8sClient, tfExecutionJob); err != nil {
+	if err := meta.UpdateTerraformJobIfNeeded(ctx, k8sClient, tfExecutionJob); err != nil {
 		klog.ErrorS(err, types.ErrUpdateTerraformApplyJob, "Name", meta.ApplyJobName)
 		return errors.Wrap(err, types.ErrUpdateTerraformApplyJob)
 	}
 
 	if !meta.EnvChanged && tfExecutionJob.Status.Succeeded == int32(1) {
-		if err := meta.updateApplyStatus(ctx, k8sClient, types.Available, types.MessageCloudResourceDeployed); err != nil {
+		if err := meta.UpdateApplyStatus(ctx, k8sClient, types.Available, types.MessageCloudResourceDeployed); err != nil {
 			return err
 		}
 	} else {
@@ -399,7 +254,7 @@ func (r *ConfigurationReconciler) terraformApply(ctx context.Context, configurat
 		// If the state is types.InvalidRegion, no need to continue checking
 		if configuration.Status.Apply.State != types.ConfigurationProvisioningAndChecking &&
 			configuration.Status.Apply.State != types.InvalidRegion {
-			if err := meta.updateApplyStatus(ctx, r.Client, types.ConfigurationProvisioningAndChecking, types.MessageCloudResourceProvisioningAndChecking); err != nil {
+			if err := meta.UpdateApplyStatus(ctx, r.Client, types.ConfigurationProvisioningAndChecking, types.MessageCloudResourceProvisioningAndChecking); err != nil {
 				return err
 			}
 		}
@@ -407,7 +262,7 @@ func (r *ConfigurationReconciler) terraformApply(ctx context.Context, configurat
 	return nil
 }
 
-func (r *ConfigurationReconciler) terraformDestroy(ctx context.Context, configuration v1beta2.Configuration, meta *TFConfigurationMeta) error {
+func (r *ConfigurationReconciler) terraformDestroy(ctx context.Context, configuration v1beta2.Configuration, meta *process.TFConfigurationMeta) error {
 	var (
 		destroyJob batchv1.Job
 		k8sClient  = r.Client
@@ -436,20 +291,20 @@ func (r *ConfigurationReconciler) terraformDestroy(ctx context.Context, configur
 		if err := k8sClient.Get(ctx, client.ObjectKey{Name: meta.DestroyJobName, Namespace: meta.ControllerNamespace}, &destroyJob); err != nil {
 			if kerrors.IsNotFound(err) {
 				if err := r.Client.Get(ctx, client.ObjectKey{Name: configuration.Name, Namespace: configuration.Namespace}, &v1beta2.Configuration{}); err == nil {
-					if err = meta.assembleAndTriggerJob(ctx, k8sClient, TerraformDestroy); err != nil {
+					if err = meta.AssembleAndTriggerJob(ctx, k8sClient, process.TerraformDestroy); err != nil {
 						return err
 					}
 				}
 			}
 		}
-		if err := meta.updateTerraformJobIfNeeded(ctx, k8sClient, destroyJob); err != nil {
+		if err := meta.UpdateTerraformJobIfNeeded(ctx, k8sClient, destroyJob); err != nil {
 			klog.ErrorS(err, types.ErrUpdateTerraformApplyJob, "Name", meta.DestroyJobName)
 			return errors.Wrap(err, types.ErrUpdateTerraformApplyJob)
 		}
 	}
 
 	// destroying
-	if err := meta.updateDestroyStatus(ctx, k8sClient, types.ConfigurationDestroying, types.MessageCloudResourceDestroying); err != nil {
+	if err := meta.UpdateDestroyStatus(ctx, k8sClient, types.ConfigurationDestroying, types.MessageCloudResourceDestroying); err != nil {
 		return err
 	}
 
@@ -475,7 +330,7 @@ func (r *ConfigurationReconciler) terraformDestroy(ctx context.Context, configur
 	return errors.New(types.MessageDestroyJobNotCompleted)
 }
 
-func (r *ConfigurationReconciler) cleanUpSubResources(ctx context.Context, configuration v1beta2.Configuration, meta *TFConfigurationMeta) error {
+func (r *ConfigurationReconciler) cleanUpSubResources(ctx context.Context, configuration v1beta2.Configuration, meta *process.TFConfigurationMeta) error {
 	var k8sClient = r.Client
 
 	// 1. delete connectionSecret
@@ -488,7 +343,7 @@ func (r *ConfigurationReconciler) cleanUpSubResources(ctx context.Context, confi
 	}
 
 	// 2~5. delete the jobs, variable secrets, configuration configmaps
-	type cleanupResourceFunc func(ctx context.Context, meta *TFConfigurationMeta, k8sClient client.Client) error
+	type cleanupResourceFunc func(ctx context.Context, meta *process.TFConfigurationMeta, k8sClient client.Client) error
 	resourceToCleanup := []cleanupResourceFunc{
 		deleteApplyJob,
 		deleteDestroyJob,
@@ -511,7 +366,7 @@ func (r *ConfigurationReconciler) cleanUpSubResources(ctx context.Context, confi
 	return nil
 }
 
-func (r *ConfigurationReconciler) preCheckResourcesSetting(meta *TFConfigurationMeta) error {
+func (r *ConfigurationReconciler) preCheckResourcesSetting(meta *process.TFConfigurationMeta) error {
 
 	meta.ResourceQuota.ResourcesLimitsCPU = os.Getenv("RESOURCES_LIMITS_CPU")
 	if meta.ResourceQuota.ResourcesLimitsCPU != "" {
@@ -556,7 +411,7 @@ func (r *ConfigurationReconciler) preCheckResourcesSetting(meta *TFConfiguration
 	return nil
 }
 
-func (r *ConfigurationReconciler) preCheck(ctx context.Context, configuration *v1beta2.Configuration, meta *TFConfigurationMeta) error {
+func (r *ConfigurationReconciler) preCheck(ctx context.Context, configuration *v1beta2.Configuration, meta *process.TFConfigurationMeta) error {
 	var k8sClient = r.Client
 
 	meta.TerraformImage = os.Getenv("TERRAFORM_IMAGE")
@@ -587,7 +442,7 @@ func (r *ConfigurationReconciler) preCheck(ctx context.Context, configuration *v
 	// Validation: 1) validate Configuration itself
 	configurationType, err := tfcfg.ValidConfigurationObject(configuration)
 	if err != nil {
-		if updateErr := meta.updateApplyStatus(ctx, k8sClient, types.ConfigurationStaticCheckFailed, err.Error()); updateErr != nil {
+		if updateErr := meta.UpdateApplyStatus(ctx, k8sClient, types.ConfigurationStaticCheckFailed, err.Error()); updateErr != nil {
 			return updateErr
 		}
 		return err
@@ -602,7 +457,7 @@ func (r *ConfigurationReconciler) preCheck(ctx context.Context, configuration *v
 			if err != nil {
 				msg = err.Error()
 			}
-			if updateStatusErr := meta.updateApplyStatus(ctx, k8sClient, types.Authorizing, msg); updateStatusErr != nil {
+			if updateStatusErr := meta.UpdateApplyStatus(ctx, k8sClient, types.Authorizing, msg); updateStatusErr != nil {
 				return errors.Wrap(updateStatusErr, msg)
 			}
 			return errors.New(msg)
@@ -614,7 +469,7 @@ func (r *ConfigurationReconciler) preCheck(ctx context.Context, configuration *v
 			}
 			meta.JobEnv = jobEnv
 		}
-		if err := meta.getCredentials(ctx, k8sClient, p); err != nil {
+		if err := meta.GetCredentials(ctx, k8sClient, p); err != nil {
 			return err
 		}
 	}
@@ -624,7 +479,7 @@ func (r *ConfigurationReconciler) preCheck(ctx context.Context, configuration *v
 	   2. TerraformCredentialsSecretReference
 	   3. TerraformRCConfigMapReference
 	   4. TerraformCredentialsHelperConfigMapReference*/
-	if err := meta.validateSecretAndConfigMap(ctx, k8sClient); err != nil {
+	if err := meta.ValidateSecretAndConfigMap(ctx, k8sClient); err != nil {
 		return err
 	}
 
@@ -641,18 +496,18 @@ func (r *ConfigurationReconciler) preCheck(ctx context.Context, configuration *v
 	}
 
 	if configuration.ObjectMeta.DeletionTimestamp.IsZero() {
-		if err := meta.storeTFConfiguration(ctx, k8sClient); err != nil {
+		if err := meta.StoreTFConfiguration(ctx, k8sClient); err != nil {
 			return err
 		}
 	}
 
 	if meta.ConfigurationChanged {
 		klog.InfoS("Configuration hanged, reloading...")
-		return meta.updateApplyStatus(ctx, k8sClient, types.ConfigurationReloading, types.ConfigurationReloadingAsHCLChanged)
+		return meta.UpdateApplyStatus(ctx, k8sClient, types.ConfigurationReloading, types.ConfigurationReloadingAsHCLChanged)
 	}
 
 	// Check whether env changes
-	if err := meta.prepareTFVariables(configuration); err != nil {
+	if err := meta.PrepareTFVariables(configuration); err != nil {
 		return err
 	}
 
@@ -677,7 +532,7 @@ func (r *ConfigurationReconciler) preCheck(ctx context.Context, configuration *v
 			if val, ok := variableInSecret.Data[k]; !ok || !bytes.Equal(v, val) {
 				meta.EnvChanged = true
 				klog.Info("Job's env changed")
-				if err := meta.updateApplyStatus(ctx, k8sClient, types.ConfigurationReloading, types.ConfigurationReloadingAsVariableChanged); err != nil {
+				if err := meta.UpdateApplyStatus(ctx, k8sClient, types.ConfigurationReloading, types.ConfigurationReloadingAsVariableChanged); err != nil {
 					return err
 				}
 				break
@@ -687,600 +542,7 @@ func (r *ConfigurationReconciler) preCheck(ctx context.Context, configuration *v
 		return err
 	}
 
-	return createTerraformExecutorClusterRole(ctx, k8sClient, fmt.Sprintf("%s-%s", meta.ControllerNamespace, ClusterRoleName))
-}
-
-func (meta *TFConfigurationMeta) validateSecretAndConfigMap(ctx context.Context, k8sClient client.Client) error {
-
-	secretConfigMapToCheck := []struct {
-		ref           *v1.SecretReference
-		notFoundState types.ConfigurationState
-		isSecret      bool
-		neededKeys    []string
-		errKey        string
-	}{
-		{
-			ref:           meta.GitCredentialsSecretReference,
-			notFoundState: types.InvalidGitCredentialsSecretReference,
-			isSecret:      true,
-			neededKeys:    []string{GitCredsKnownHosts, v1.SSHAuthPrivateKey},
-			errKey:        "git credentials",
-		},
-		{
-			ref:           meta.TerraformCredentialsSecretReference,
-			notFoundState: types.InvalidTerraformCredentialsSecretReference,
-			isSecret:      true,
-			neededKeys:    []string{TerraformCredentials},
-			errKey:        "terraform credentials",
-		},
-		{
-			ref:           meta.TerraformRCConfigMapReference,
-			notFoundState: types.InvalidTerraformRCConfigMapReference,
-			isSecret:      false,
-			neededKeys:    []string{TerraformRegistryConfig},
-			errKey:        "terraformrc configuration",
-		},
-		{
-			ref:           meta.TerraformCredentialsHelperConfigMapReference,
-			notFoundState: types.InvalidTerraformCredentialsHelperConfigMapReference,
-			isSecret:      false,
-			neededKeys:    []string{},
-			errKey:        "terraform credentials helper",
-		},
-	}
-	for _, check := range secretConfigMapToCheck {
-		if check.ref != nil {
-			var object metav1.Object
-			var err error
-			object, err = GetSecretOrConfigMap(ctx, k8sClient, check.isSecret, check.ref, check.neededKeys, check.errKey)
-			if object == nil {
-				msg := string(check.notFoundState)
-				if err != nil {
-					msg = err.Error()
-				}
-				if updateStatusErr := meta.updateApplyStatus(ctx, k8sClient, check.notFoundState, msg); updateStatusErr != nil {
-					return errors.Wrap(updateStatusErr, msg)
-				}
-				return errors.New(msg)
-			}
-			// fix: The configmap or secret that the pod restricts from mounting must be in the same namespace as the pod,
-			//      otherwise the volume mount will fail.
-			if object.GetNamespace() != meta.ControllerNamespace {
-				objectKind := "ConfigMap"
-				if check.isSecret {
-					objectKind = "Secret"
-				}
-				msg := fmt.Sprintf("Invalid %s '%s/%s', whose namespace '%s' is different from the Configuration, cannot mount the volume,"+
-					" you can fix this issue by creating the Secret/ConfigMap in the '%s' namespace.",
-					objectKind, object.GetNamespace(), object.GetName(), meta.ControllerNamespace, meta.ControllerNamespace)
-				if updateStatusErr := meta.updateApplyStatus(ctx, k8sClient, check.notFoundState, msg); updateStatusErr != nil {
-					return errors.Wrap(updateStatusErr, msg)
-				}
-				return errors.New(msg)
-			}
-		}
-	}
-	return nil
-}
-
-func (meta *TFConfigurationMeta) updateApplyStatus(ctx context.Context, k8sClient client.Client, state types.ConfigurationState, message string) error {
-	var configuration v1beta2.Configuration
-	if err := k8sClient.Get(ctx, client.ObjectKey{Name: meta.Name, Namespace: meta.Namespace}, &configuration); err == nil {
-		configuration.Status.Apply = v1beta2.ConfigurationApplyStatus{
-			State:   state,
-			Message: message,
-			Region:  meta.Region,
-		}
-		configuration.Status.ObservedGeneration = configuration.Generation
-		if state == types.Available {
-			outputs, err := meta.getTFOutputs(ctx, k8sClient, configuration)
-			if err != nil {
-				klog.InfoS("Failed to get outputs", "error", err)
-				configuration.Status.Apply = v1beta2.ConfigurationApplyStatus{
-					State:   types.GeneratingOutputs,
-					Message: types.ErrGenerateOutputs + ": " + err.Error(),
-				}
-			} else {
-				configuration.Status.Apply.Outputs = outputs
-			}
-		}
-
-		return k8sClient.Status().Update(ctx, &configuration)
-	}
-	return nil
-}
-
-func (meta *TFConfigurationMeta) updateDestroyStatus(ctx context.Context, k8sClient client.Client, state types.ConfigurationState, message string) error {
-	var configuration v1beta2.Configuration
-	if err := k8sClient.Get(ctx, client.ObjectKey{Name: meta.Name, Namespace: meta.Namespace}, &configuration); err == nil {
-		configuration.Status.Destroy = v1beta2.ConfigurationDestroyStatus{
-			State:   state,
-			Message: message,
-		}
-		return k8sClient.Status().Update(ctx, &configuration)
-	}
-	return nil
-}
-
-func (meta *TFConfigurationMeta) assembleAndTriggerJob(ctx context.Context, k8sClient client.Client, executionType TerraformExecutionType) error {
-	// apply rbac
-	if err := createTerraformExecutorServiceAccount(ctx, k8sClient, meta.ControllerNamespace, ServiceAccountName); err != nil {
-		return err
-	}
-	if err := createTerraformExecutorClusterRoleBinding(ctx, k8sClient, meta.ControllerNamespace, fmt.Sprintf("%s-%s", meta.ControllerNamespace, ClusterRoleName), ServiceAccountName); err != nil {
-		return err
-	}
-
-	job := meta.assembleTerraformJob(executionType)
-
-	return k8sClient.Create(ctx, job)
-}
-
-// updateTerraformJobIfNeeded will set deletion finalizer to the Terraform job if its envs are changed, which will result in
-// deleting the job. Finally, a new Terraform job will be generated
-func (meta *TFConfigurationMeta) updateTerraformJobIfNeeded(ctx context.Context, k8sClient client.Client, job batchv1.Job) error {
-	// if either one changes, delete the job
-	if meta.EnvChanged || meta.ConfigurationChanged {
-		klog.InfoS("about to delete job", "Name", job.Name, "Namespace", job.Namespace)
-		var j batchv1.Job
-		if err := k8sClient.Get(ctx, client.ObjectKey{Name: job.Name, Namespace: job.Namespace}, &j); err == nil {
-			if deleteErr := k8sClient.Delete(ctx, &job, client.PropagationPolicy(metav1.DeletePropagationBackground)); deleteErr != nil {
-				return deleteErr
-			}
-		}
-		var s v1.Secret
-		if err := k8sClient.Get(ctx, client.ObjectKey{Name: meta.VariableSecretName, Namespace: meta.ControllerNamespace}, &s); err == nil {
-			if deleteErr := k8sClient.Delete(ctx, &s); deleteErr != nil {
-				return deleteErr
-			}
-		}
-	}
-	return nil
-}
-
-func (meta *TFConfigurationMeta) assembleTerraformJob(executionType TerraformExecutionType) *batchv1.Job {
-	var (
-		initContainer           v1.Container
-		tfPreApplyInitContainer v1.Container
-		initContainers          []v1.Container
-		parallelism             int32 = 1
-		completions             int32 = 1
-	)
-
-	executorVolumes := meta.assembleExecutorVolumes()
-	initContainerVolumeMounts := []v1.VolumeMount{
-		{
-			Name:      meta.Name,
-			MountPath: WorkingVolumeMountPath,
-		},
-		{
-			Name:      InputTFConfigurationVolumeName,
-			MountPath: InputTFConfigurationVolumeMountPath,
-		},
-		{
-			Name:      BackendVolumeName,
-			MountPath: BackendVolumeMountPath,
-		},
-	}
-
-	if meta.TerraformCredentialsSecretReference != nil {
-		initContainerVolumeMounts = append(initContainerVolumeMounts,
-			v1.VolumeMount{
-				Name:      TerraformCredentialsConfigVolumeName,
-				MountPath: TerraformCredentialsConfigVolumeMountPath,
-			})
-	}
-
-	if meta.TerraformRCConfigMapReference != nil {
-		initContainerVolumeMounts = append(initContainerVolumeMounts,
-			v1.VolumeMount{
-				Name:      TerraformRCConfigVolumeName,
-				MountPath: TerraformRCConfigVolumeMountPath,
-			})
-	}
-
-	if meta.TerraformCredentialsHelperConfigMapReference != nil {
-		initContainerVolumeMounts = append(initContainerVolumeMounts,
-			v1.VolumeMount{
-				Name:      TerraformCredentialsHelperConfigVolumeName,
-				MountPath: TerraformCredentialsHelperConfigVolumeMountPath,
-			})
-	}
-
-	// prepare local Terraform .tf files
-	initContainer = v1.Container{
-		Name:            "prepare-input-terraform-configurations",
-		Image:           meta.BusyboxImage,
-		ImagePullPolicy: v1.PullIfNotPresent,
-		Command: []string{
-			"sh",
-			"-c",
-			fmt.Sprintf("cp %s/* %s", InputTFConfigurationVolumeMountPath, WorkingVolumeMountPath),
-		},
-		VolumeMounts: initContainerVolumeMounts,
-	}
-
-	initContainers = append(initContainers, initContainer)
-
-	hclPath := filepath.Join(BackendVolumeMountPath, meta.RemoteGitPath)
-
-	if meta.RemoteGit != "" {
-		cloneCommand := fmt.Sprintf("git clone %s %s && cp -r %s/* %s", meta.RemoteGit, BackendVolumeMountPath, hclPath, WorkingVolumeMountPath)
-
-		// Check for git credentials, mount the SSH known hosts and private key, add private key into the SSH authentication agent
-		if meta.GitCredentialsSecretReference != nil {
-			initContainerVolumeMounts = append(initContainerVolumeMounts,
-				v1.VolumeMount{
-					Name:      GitAuthConfigVolumeName,
-					MountPath: GitAuthConfigVolumeMountPath,
-				})
-
-			sshCommand := fmt.Sprintf("eval `ssh-agent` && ssh-add %s/%s", GitAuthConfigVolumeMountPath, v1.SSHAuthPrivateKey)
-			cloneCommand = fmt.Sprintf("%s && %s", sshCommand, cloneCommand)
-		}
-
-		command := []string{
-			"sh",
-			"-c",
-			cloneCommand,
-		}
-
-		initContainers = append(initContainers,
-			v1.Container{
-				Name:            "git-configuration",
-				Image:           meta.GitImage,
-				ImagePullPolicy: v1.PullIfNotPresent,
-				Command:         command,
-				VolumeMounts:    initContainerVolumeMounts,
-			})
-	}
-
-	// run `terraform init`
-	tfPreApplyInitContainer = v1.Container{
-		Name:            terraformInitContainerName,
-		Image:           meta.TerraformImage,
-		ImagePullPolicy: v1.PullIfNotPresent,
-		Command: []string{
-			"sh",
-			"-c",
-			"terraform init",
-		},
-		VolumeMounts: initContainerVolumeMounts,
-		Env:          meta.Envs,
-	}
-	initContainers = append(initContainers, tfPreApplyInitContainer)
-
-	container := v1.Container{
-		Name:            terraformContainerName,
-		Image:           meta.TerraformImage,
-		ImagePullPolicy: v1.PullIfNotPresent,
-		Command: []string{
-			"bash",
-			"-c",
-			fmt.Sprintf("terraform %s -lock=false -auto-approve", executionType),
-		},
-		VolumeMounts: []v1.VolumeMount{
-			{
-				Name:      meta.Name,
-				MountPath: WorkingVolumeMountPath,
-			},
-			{
-				Name:      InputTFConfigurationVolumeName,
-				MountPath: InputTFConfigurationVolumeMountPath,
-			},
-		},
-		Env: meta.Envs,
-	}
-
-	if meta.ResourceQuota.ResourcesLimitsCPU != "" || meta.ResourceQuota.ResourcesLimitsMemory != "" ||
-		meta.ResourceQuota.ResourcesRequestsCPU != "" || meta.ResourceQuota.ResourcesRequestsMemory != "" {
-		resourceRequirements := v1.ResourceRequirements{}
-		if meta.ResourceQuota.ResourcesLimitsCPU != "" || meta.ResourceQuota.ResourcesLimitsMemory != "" {
-			resourceRequirements.Limits = map[v1.ResourceName]resource.Quantity{}
-			if meta.ResourceQuota.ResourcesLimitsCPU != "" {
-				resourceRequirements.Limits["cpu"] = meta.ResourceQuota.ResourcesLimitsCPUQuantity
-			}
-			if meta.ResourceQuota.ResourcesLimitsMemory != "" {
-				resourceRequirements.Limits["memory"] = meta.ResourceQuota.ResourcesLimitsMemoryQuantity
-			}
-		}
-		if meta.ResourceQuota.ResourcesRequestsCPU != "" || meta.ResourceQuota.ResourcesLimitsMemory != "" {
-			resourceRequirements.Requests = map[v1.ResourceName]resource.Quantity{}
-			if meta.ResourceQuota.ResourcesRequestsCPU != "" {
-				resourceRequirements.Requests["cpu"] = meta.ResourceQuota.ResourcesRequestsCPUQuantity
-			}
-			if meta.ResourceQuota.ResourcesRequestsMemory != "" {
-				resourceRequirements.Requests["memory"] = meta.ResourceQuota.ResourcesRequestsMemoryQuantity
-			}
-		}
-		container.Resources = resourceRequirements
-	}
-
-	name := meta.ApplyJobName
-	if executionType == TerraformDestroy {
-		name = meta.DestroyJobName
-	}
-
-	return &batchv1.Job{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Job",
-			APIVersion: "batch/v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: meta.ControllerNamespace,
-		},
-		Spec: batchv1.JobSpec{
-			Parallelism:  &parallelism,
-			Completions:  &completions,
-			BackoffLimit: &meta.BackoffLimit,
-			Template: v1.PodTemplateSpec{
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{
-						// This annotation will prevent istio-proxy sidecar injection in the pods
-						// as having the sidecar would have kept the Job in `Running` state and would
-						// not transition to `Completed`
-						"sidecar.istio.io/inject": "false",
-					},
-				},
-				Spec: v1.PodSpec{
-					// InitContainer will copy Terraform configuration files to working directory and create Terraform
-					// state file directory in advance
-					InitContainers: initContainers,
-					// Container terraform-executor will first copy predefined terraform.d to working directory, and
-					// then run terraform init/apply.
-					Containers:         []v1.Container{container},
-					ServiceAccountName: ServiceAccountName,
-					Volumes:            executorVolumes,
-					RestartPolicy:      v1.RestartPolicyOnFailure,
-					NodeSelector:       meta.JobNodeSelector,
-				},
-			},
-		},
-	}
-}
-
-func (meta *TFConfigurationMeta) assembleExecutorVolumes() []v1.Volume {
-	workingVolume := v1.Volume{Name: meta.Name}
-	workingVolume.EmptyDir = &v1.EmptyDirVolumeSource{}
-	inputTFConfigurationVolume := meta.createConfigurationVolume()
-	tfBackendVolume := meta.createTFBackendVolume()
-	executorVolumes := []v1.Volume{workingVolume, inputTFConfigurationVolume, tfBackendVolume}
-	secretOrConfigMapReferences := []struct {
-		ref        *v1.SecretReference
-		volumeName string
-		isSecret   bool
-	}{
-		{
-			ref:        meta.GitCredentialsSecretReference,
-			volumeName: GitAuthConfigVolumeName,
-			isSecret:   true,
-		},
-		{
-			ref:        meta.TerraformCredentialsSecretReference,
-			volumeName: TerraformCredentialsConfigVolumeName,
-			isSecret:   true,
-		},
-		{
-			ref:        meta.TerraformRCConfigMapReference,
-			volumeName: TerraformRCConfigVolumeName,
-			isSecret:   false,
-		},
-		{
-			ref:        meta.TerraformCredentialsHelperConfigMapReference,
-			volumeName: TerraformCredentialsHelperConfigVolumeName,
-			isSecret:   false,
-		},
-	}
-	for _, ref := range secretOrConfigMapReferences {
-		if ref.ref != nil {
-			executorVolumes = append(executorVolumes, meta.createSecretOrConfigMapVolume(ref.isSecret, ref.ref.Name, ref.volumeName))
-		}
-	}
-	return executorVolumes
-}
-
-func (meta *TFConfigurationMeta) createConfigurationVolume() v1.Volume {
-	inputCMVolumeSource := v1.ConfigMapVolumeSource{}
-	inputCMVolumeSource.Name = meta.ConfigurationCMName
-	inputTFConfigurationVolume := v1.Volume{Name: InputTFConfigurationVolumeName}
-	inputTFConfigurationVolume.ConfigMap = &inputCMVolumeSource
-	return inputTFConfigurationVolume
-
-}
-
-func (meta *TFConfigurationMeta) createTFBackendVolume() v1.Volume {
-	gitVolume := v1.Volume{Name: BackendVolumeName}
-	gitVolume.EmptyDir = &v1.EmptyDirVolumeSource{}
-	return gitVolume
-}
-
-func (meta *TFConfigurationMeta) createSecretOrConfigMapVolume(isSecret bool, secretOrConfigMapReferenceName string, volumeName string) v1.Volume {
-	var defaultMode int32 = 0400
-	volume := v1.Volume{Name: volumeName}
-	if isSecret {
-		volumeSource := v1.SecretVolumeSource{}
-		volumeSource.SecretName = secretOrConfigMapReferenceName
-		volumeSource.DefaultMode = &defaultMode
-		volume.Secret = &volumeSource
-	} else {
-		volumeSource := v1.ConfigMapVolumeSource{}
-		volumeSource.Name = secretOrConfigMapReferenceName
-		volumeSource.DefaultMode = &defaultMode
-		volume.ConfigMap = &volumeSource
-	}
-	return volume
-}
-
-// TfStateProperty is the tf state property for an output
-type TfStateProperty struct {
-	Value interface{} `json:"value,omitempty"`
-	Type  interface{} `json:"type,omitempty"`
-}
-
-// ToProperty converts TfStateProperty type to Property
-func (tp *TfStateProperty) ToProperty() (v1beta2.Property, error) {
-	var (
-		property v1beta2.Property
-		err      error
-	)
-	sv, err := tfcfg.Interface2String(tp.Value)
-	if err != nil {
-		return property, errors.Wrapf(err, "failed to convert value %s of terraform state outputs to string", tp.Value)
-	}
-	property = v1beta2.Property{
-		Value: sv,
-	}
-	return property, err
-}
-
-// TFState is Terraform State
-type TFState struct {
-	Outputs map[string]TfStateProperty `json:"outputs"`
-}
-
-func (meta *TFConfigurationMeta) isTFStateGenerated(ctx context.Context) bool {
-	// 1. exist backend
-	if meta.Backend == nil {
-		return false
-	}
-	// 2. and exist tfstate file
-	_, err := meta.Backend.GetTFStateJSON(ctx)
-	return err == nil
-}
-
-//nolint:funlen
-func (meta *TFConfigurationMeta) getTFOutputs(ctx context.Context, k8sClient client.Client, configuration v1beta2.Configuration) (map[string]v1beta2.Property, error) {
-	var tfStateJSON []byte
-	var err error
-	if meta.Backend != nil {
-		tfStateJSON, err = meta.Backend.GetTFStateJSON(ctx)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	var tfState TFState
-	if err := json.Unmarshal(tfStateJSON, &tfState); err != nil {
-		return nil, err
-	}
-	outputs := make(map[string]v1beta2.Property)
-	for k, v := range tfState.Outputs {
-		property, err := v.ToProperty()
-		if err != nil {
-			return outputs, err
-		}
-		outputs[k] = property
-	}
-	writeConnectionSecretToReference := configuration.Spec.WriteConnectionSecretToReference
-	if writeConnectionSecretToReference == nil || writeConnectionSecretToReference.Name == "" {
-		return outputs, nil
-	}
-
-	name := writeConnectionSecretToReference.Name
-	ns := writeConnectionSecretToReference.Namespace
-	if ns == "" {
-		ns = defaultNamespace
-	}
-	data := make(map[string][]byte)
-	for k, v := range outputs {
-		data[k] = []byte(v.Value)
-	}
-	var gotSecret v1.Secret
-	configurationName := configuration.ObjectMeta.Name
-	if err := k8sClient.Get(ctx, client.ObjectKey{Name: name, Namespace: ns}, &gotSecret); err != nil {
-		if kerrors.IsNotFound(err) {
-			var secret = v1.Secret{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      name,
-					Namespace: ns,
-					Labels: map[string]string{
-						"terraform.core.oam.dev/created-by":      "terraform-controller",
-						"terraform.core.oam.dev/owned-by":        configurationName,
-						"terraform.core.oam.dev/owned-namespace": configuration.Namespace,
-					},
-				},
-				TypeMeta: metav1.TypeMeta{Kind: "Secret"},
-				Data:     data,
-			}
-			err = k8sClient.Create(ctx, &secret)
-			if kerrors.IsAlreadyExists(err) {
-				return nil, fmt.Errorf("secret(%s) already exists", name)
-			} else if err != nil {
-				return nil, err
-			}
-		}
-	} else {
-		// check the owner of this secret
-		labels := gotSecret.ObjectMeta.Labels
-		ownerName := labels["terraform.core.oam.dev/owned-by"]
-		ownerNamespace := labels["terraform.core.oam.dev/owned-namespace"]
-		if (ownerName != "" && ownerName != configurationName) ||
-			(ownerNamespace != "" && ownerNamespace != configuration.Namespace) {
-			errMsg := fmt.Sprintf(
-				"configuration(namespace: %s ; name: %s) cannot update secret(namespace: %s ; name: %s) whose owner is configuration(namespace: %s ; name: %s)",
-				configuration.Namespace, configurationName,
-				gotSecret.Namespace, name,
-				ownerNamespace, ownerName,
-			)
-			klog.ErrorS(err, "fail to update backend secret")
-			return nil, errors.New(errMsg)
-		}
-		gotSecret.Data = data
-		if err := k8sClient.Update(ctx, &gotSecret); err != nil {
-			return nil, err
-		}
-	}
-	return outputs, nil
-}
-
-func (meta *TFConfigurationMeta) prepareTFVariables(configuration *v1beta2.Configuration) error {
-	var (
-		envs []v1.EnvVar
-		data = map[string][]byte{}
-	)
-
-	if configuration == nil {
-		return errors.New("configuration is nil")
-	}
-	if !configuration.Spec.InlineCredentials && meta.ProviderReference == nil {
-		return errors.New("The referenced provider could not be retrieved")
-	}
-
-	tfVariable, err := getTerraformJSONVariable(configuration.Spec.Variable)
-	if err != nil {
-		return errors.Wrap(err, fmt.Sprintf("failed to get Terraform JSON variables from Configuration Variables %v", configuration.Spec.Variable))
-	}
-	for k, v := range tfVariable {
-		envValue, err := tfcfg.Interface2String(v)
-		if err != nil {
-			return err
-		}
-		data[k] = []byte(envValue)
-	}
-
-	if !configuration.Spec.InlineCredentials && meta.Credentials == nil {
-		return errors.New(provider.ErrCredentialNotRetrieved)
-	}
-	for k, v := range meta.Credentials {
-		data[k] = []byte(v)
-	}
-	for k, v := range meta.JobEnv {
-		envValue, err := tfcfg.Interface2String(v)
-		if err != nil {
-			return err
-		}
-		data[k] = []byte(envValue)
-	}
-	for k := range data {
-		valueFrom := &v1.EnvVarSource{SecretKeyRef: &v1.SecretKeySelector{Key: k}}
-		valueFrom.SecretKeyRef.Name = meta.VariableSecretName
-		envs = append(envs, v1.EnvVar{Name: k, ValueFrom: valueFrom})
-	}
-	meta.Envs = envs
-	meta.VariableSecretData = data
-
-	return nil
+	return util.CreateTerraformExecutorClusterRole(ctx, k8sClient, fmt.Sprintf("%s-%s", meta.ControllerNamespace, ClusterRoleName))
 }
 
 // SetupWithManager setups with a manager
@@ -1290,20 +552,7 @@ func (r *ConfigurationReconciler) SetupWithManager(mgr ctrl.Manager) error {
 		Complete(r)
 }
 
-func getTerraformJSONVariable(tfVariables *runtime.RawExtension) (map[string]interface{}, error) {
-	variables, err := tfcfg.RawExtension2Map(tfVariables)
-	if err != nil {
-		return nil, err
-	}
-	var environments = make(map[string]interface{})
-
-	for k, v := range variables {
-		environments[fmt.Sprintf("TF_VAR_%s", k)] = v
-	}
-	return environments, nil
-}
-
-func deleteConfigMap(ctx context.Context, meta *TFConfigurationMeta, k8sClient client.Client) error {
+func deleteConfigMap(ctx context.Context, meta *process.TFConfigurationMeta, k8sClient client.Client) error {
 	var cm v1.ConfigMap
 	// We have four cases when upgrading. There are three combinations of name and namespace.
 	// TODO compatible for case 4
@@ -1327,7 +576,7 @@ func deleteConfigMap(ctx context.Context, meta *TFConfigurationMeta, k8sClient c
 	return nil
 }
 
-func deleteVariableSecret(ctx context.Context, meta *TFConfigurationMeta, k8sClient client.Client) error {
+func deleteVariableSecret(ctx context.Context, meta *process.TFConfigurationMeta, k8sClient client.Client) error {
 	var variableSecret v1.Secret
 	// see TFConfigurationMeta.deleteConfigMap
 	possibleCombination := [][2]string{
@@ -1346,7 +595,7 @@ func deleteVariableSecret(ctx context.Context, meta *TFConfigurationMeta, k8sCli
 	return nil
 }
 
-func deleteApplyJob(ctx context.Context, meta *TFConfigurationMeta, k8sClient client.Client) error {
+func deleteApplyJob(ctx context.Context, meta *process.TFConfigurationMeta, k8sClient client.Client) error {
 	var job batchv1.Job
 	// see TFConfigurationMeta.deleteConfigMap
 	possibleCombination := [][2]string{
@@ -1365,7 +614,7 @@ func deleteApplyJob(ctx context.Context, meta *TFConfigurationMeta, k8sClient cl
 	return nil
 }
 
-func deleteDestroyJob(ctx context.Context, meta *TFConfigurationMeta, k8sClient client.Client) error {
+func deleteDestroyJob(ctx context.Context, meta *process.TFConfigurationMeta, k8sClient client.Client) error {
 	var job batchv1.Job
 	// see TFConfigurationMeta.deleteConfigMap
 	possibleCombination := [][2]string{
@@ -1397,176 +646,4 @@ func deleteConnectionSecret(ctx context.Context, k8sClient client.Client, name, 
 		return k8sClient.Delete(ctx, &connectionSecret)
 	}
 	return nil
-}
-
-func (meta *TFConfigurationMeta) createOrUpdateConfigMap(ctx context.Context, k8sClient client.Client, data map[string]string) error {
-	var gotCM v1.ConfigMap
-	if err := k8sClient.Get(ctx, client.ObjectKey{Name: meta.ConfigurationCMName, Namespace: meta.ControllerNamespace}, &gotCM); err != nil {
-		if !kerrors.IsNotFound(err) {
-			return err
-		}
-		cm := v1.ConfigMap{
-			TypeMeta: metav1.TypeMeta{APIVersion: "v1", Kind: "ConfigMap"},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      meta.ConfigurationCMName,
-				Namespace: meta.ControllerNamespace,
-			},
-			Data: data,
-		}
-
-		if err := k8sClient.Create(ctx, &cm); err != nil {
-			return errors.Wrap(err, "failed to create TF configuration ConfigMap")
-		}
-
-		return nil
-	}
-
-	if !reflect.DeepEqual(gotCM.Data, data) {
-		gotCM.Data = data
-
-		return errors.Wrap(k8sClient.Update(ctx, &gotCM), "failed to update TF configuration ConfigMap")
-	}
-
-	return nil
-}
-
-func (meta *TFConfigurationMeta) prepareTFInputConfigurationData() map[string]string {
-	var dataName string
-	switch meta.ConfigurationType {
-	case types.ConfigurationHCL:
-		dataName = types.TerraformHCLConfigurationName
-	case types.ConfigurationRemote:
-		dataName = "terraform-backend.tf"
-	}
-	data := map[string]string{dataName: meta.CompleteConfiguration, "kubeconfig": ""}
-	return data
-}
-
-// storeTFConfiguration will store Terraform configuration to ConfigMap
-func (meta *TFConfigurationMeta) storeTFConfiguration(ctx context.Context, k8sClient client.Client) error {
-	data := meta.prepareTFInputConfigurationData()
-	return meta.createOrUpdateConfigMap(ctx, k8sClient, data)
-}
-
-// CheckWhetherConfigurationChanges will check whether configuration is changed
-func (meta *TFConfigurationMeta) CheckWhetherConfigurationChanges(ctx context.Context, k8sClient client.Client, configurationType types.ConfigurationType) error {
-	switch configurationType {
-	case types.ConfigurationHCL:
-		var cm v1.ConfigMap
-		if err := k8sClient.Get(ctx, client.ObjectKey{Name: meta.ConfigurationCMName, Namespace: meta.ControllerNamespace}, &cm); err != nil {
-			if kerrors.IsNotFound(err) {
-				return nil
-			}
-			return err
-		}
-		meta.ConfigurationChanged = cm.Data[types.TerraformHCLConfigurationName] != meta.CompleteConfiguration
-		if meta.ConfigurationChanged {
-			klog.InfoS("Configuration HCL changed", "ConfigMap", cm.Data[types.TerraformHCLConfigurationName],
-				"RenderedCompletedConfiguration", meta.CompleteConfiguration)
-		}
-
-		return nil
-	case types.ConfigurationRemote:
-		meta.ConfigurationChanged = false
-		return nil
-	default:
-		return errors.New("unsupported configuration type, only HCL or Remote is supported")
-	}
-}
-
-// getCredentials will get credentials from secret of the Provider
-func (meta *TFConfigurationMeta) getCredentials(ctx context.Context, k8sClient client.Client, providerObj *v1beta1.Provider) error {
-	region, err := tfcfg.SetRegion(ctx, k8sClient, meta.Namespace, meta.Name, providerObj)
-	if err != nil {
-		return err
-	}
-	credentials, err := provider.GetProviderCredentials(ctx, k8sClient, providerObj, region)
-	if err != nil {
-		return err
-	}
-	if credentials == nil {
-		return errors.New(provider.ErrCredentialNotRetrieved)
-	}
-	meta.Credentials = credentials
-	meta.Region = region
-	return nil
-}
-
-func (meta *TFConfigurationMeta) KeepLegacySubResourceMetas() {
-	meta.LegacySubResources.Namespace = meta.Namespace
-	meta.LegacySubResources.ApplyJobName = meta.ApplyJobName
-	meta.LegacySubResources.DestroyJobName = meta.DestroyJobName
-	meta.LegacySubResources.ConfigurationCMName = meta.ConfigurationCMName
-	meta.LegacySubResources.VariableSecretName = meta.VariableSecretName
-}
-
-func (meta *TFConfigurationMeta) getApplyJob(ctx context.Context, k8sClient client.Client, job *batchv1.Job) error {
-	if err := k8sClient.Get(ctx, client.ObjectKey{Name: meta.LegacySubResources.ApplyJobName, Namespace: meta.LegacySubResources.Namespace}, job); err == nil {
-		klog.InfoS("Found legacy apply job", "Configuration", fmt.Sprintf("%s/%s", meta.Name, meta.Namespace),
-			"Job", fmt.Sprintf("%s/%s", meta.LegacySubResources.Namespace, meta.LegacySubResources.ApplyJobName))
-		return nil
-	}
-	err := k8sClient.Get(ctx, client.ObjectKey{Name: meta.ApplyJobName, Namespace: meta.ControllerNamespace}, job)
-	return err
-}
-
-// RenderConfiguration will compose the Terraform configuration with hcl/json and backend
-func (meta *TFConfigurationMeta) RenderConfiguration(configuration *v1beta2.Configuration, configurationType types.ConfigurationType) (string, backend.Backend, error) {
-	backendInterface, err := backend.ParseConfigurationBackend(configuration, meta.K8sClient, meta.Credentials, meta.ControllerNSSpecified)
-	if err != nil {
-		return "", nil, errors.Wrap(err, "failed to prepare Terraform backend configuration")
-	}
-
-	switch configurationType {
-	case types.ConfigurationHCL:
-		completedConfiguration := configuration.Spec.HCL
-		completedConfiguration += "\n" + backendInterface.HCL()
-		return completedConfiguration, backendInterface, nil
-	case types.ConfigurationRemote:
-		return backendInterface.HCL(), backendInterface, nil
-	default:
-		return "", nil, errors.New("Unsupported Configuration Type")
-	}
-}
-
-func GetSecretOrConfigMap(ctx context.Context, k8sClient client.Client, isSecret bool, ref *v1.SecretReference, neededKeys []string, errKey string) (metav1.Object, error) {
-	secret := &v1.Secret{}
-	configMap := &v1.ConfigMap{}
-	var err error
-	// key to determine if it is a secret or config map
-	var typeKey string
-	if isSecret {
-		namespacedName := client.ObjectKey{Name: ref.Name, Namespace: ref.Namespace}
-		err = k8sClient.Get(ctx, namespacedName, secret)
-		typeKey = "secret"
-	} else {
-		namespacedName := client.ObjectKey{Name: ref.Name, Namespace: ref.Namespace}
-		err = k8sClient.Get(ctx, namespacedName, configMap)
-		typeKey = "configmap"
-	}
-	errMsg := fmt.Sprintf("Failed to get %s %s", errKey, typeKey)
-	if err != nil {
-		klog.ErrorS(err, errMsg, "Name", ref.Name, "Namespace", ref.Namespace)
-		return nil, errors.Wrap(err, errMsg)
-	}
-	for _, key := range neededKeys {
-		var keyErr bool
-		if isSecret {
-			if _, ok := secret.Data[key]; !ok {
-				keyErr = true
-			}
-		} else {
-			if _, ok := configMap.Data[key]; !ok {
-				keyErr = true
-			}
-		}
-		if keyErr {
-			keyErr := errors.Errorf("'%s' not in %s %s", key, errKey, typeKey)
-			return nil, keyErr
-		}
-	}
-	if isSecret {
-		return secret, nil
-	}
-	return configMap, nil
 }
