+++ controllers/util/rbac.go
@@ -1,18 +1,17 @@
-package controllers
+package util
 
 import (
 	"context"
 	"fmt"
 
 	"github.com/pkg/errors"
-	v1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-func createTerraformExecutorClusterRole(ctx context.Context, k8sClient client.Client, clusterRoleName string) error {
+func CreateTerraformExecutorClusterRole(ctx context.Context, k8sClient client.Client, clusterRoleName string) error {
 	var clusterRole = rbacv1.ClusterRole{
 		TypeMeta: metav1.TypeMeta{
 			APIVersion: "rbac.authorization.k8s.io/v1",
@@ -44,7 +43,7 @@ func createTerraformExecutorClusterRole(ctx context.Context, k8sClient client.Cl
 	return nil
 }
 
-func createTerraformExecutorClusterRoleBinding(ctx context.Context, k8sClient client.Client, namespace, clusterRoleName, serviceAccountName string) error {
+func CreateTerraformExecutorClusterRoleBinding(ctx context.Context, k8sClient client.Client, namespace, clusterRoleName, serviceAccountName string) error {
 	var crbName = fmt.Sprintf("%s-tf-executor-clusterrole-binding", namespace)
 	var clusterRoleBinding = rbacv1.ClusterRoleBinding{
 		TypeMeta: metav1.TypeMeta{
@@ -72,27 +71,6 @@ func createTerraformExecutorClusterRoleBinding(ctx context.Context, k8sClient cl
 		if kerrors.IsNotFound(err) {
 			if err := k8sClient.Create(ctx, &clusterRoleBinding); err != nil {
 				return errors.Wrap(err, "failed to create ClusterRoleBinding for Terraform executor")
-			}
-		}
-	}
-	return nil
-}
-
-func createTerraformExecutorServiceAccount(ctx context.Context, k8sClient client.Client, namespace, serviceAccountName string) error {
-	var serviceAccount = v1.ServiceAccount{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: "v1",
-			Kind:       "ServiceAccount",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      serviceAccountName,
-			Namespace: namespace,
-		},
-	}
-	if err := k8sClient.Get(ctx, client.ObjectKey{Name: serviceAccountName, Namespace: namespace}, &v1.ServiceAccount{}); err != nil {
-		if kerrors.IsNotFound(err) {
-			if err := k8sClient.Create(ctx, &serviceAccount); err != nil {
-				return errors.Wrap(err, "failed to create ServiceAccount for Terraform executor")
 			}
 		}
 	}
